backend/requirements.txt
django
djangorestframework
gunicorn
psycopg2-binary
whitenoise
python-dotenv
dj-database-url
